+++ api/tasks.py
@@ -23,6 +23,7 @@ from utils.various import add_log, add_user_log
 import urllib
 import os
 import re
+import magic
 
 # TRANSCODING
 
@@ -46,8 +47,12 @@ def federate_new_sound(sound: Sound) -> int:
 
     if sound.path_artwork():
         url_artwork = url_for("get_uploads_stuff", thing="artwork_sounds", stuff=sound.path_artwork(), _external=True)
+        mime_artwork = magic.from_file(
+            os.path.join(current_app.config["UPLOADED_ARTWORKSOUNDS_DEST"], sound.path_artwork()), mime=True
+        )
+        obj_artwork = {"type": "Image", "url": url_artwork, "mediaType": mime_artwork}
     else:
-        url_artwork = None
+        obj_artwork = None
 
     licence = sound.licence_info()
     licence["id"] = str(licence["id"])  # integers makes jsonld cry
@@ -65,7 +70,7 @@ def federate_new_sound(sound: Sound) -> int:
         tags=[t.name for t in sound.tags],
         genre=sound.genre,
         licence=licence,
-        artwork=url_artwork,
+        image=obj_artwork,
         transcoded=sound.transcode_needed,
         transcode_url=(url_transcode if sound.transcode_needed else None),
     )
@@ -173,8 +178,13 @@ def create_sound_for_remote_track(activity: Activity) -> int:
     # Get user through actor
     sound.user_id = activity.actor.user.id
     sound.activity_id = activity.id
+
+    # AP standard
+    artwork = activity.payload.get("object", {}).get("image", None)
+    if artwork:
+        sound.remote_artwork_uri = artwork.get("url", None)
+
     # custom from AP
-    sound.remote_artwork_uri = activity.payload.get("object", {}).get("artwork", None)
     sound.genre = activity.payload.get("object", {}).get("genre", None)
     sound.licence = activity.payload.get("object", {}).get("licence", {}).get("id", 0)
 
@@ -698,11 +708,15 @@ def send_update_sound(sound: Sound) -> None:
 
     if sound.path_artwork():
         url_artwork = url_for("get_uploads_stuff", thing="artwork_sounds", stuff=sound.path_artwork(), _external=True)
+        mime_artwork = magic.from_file(
+            os.path.join(current_app.config["UPLOADED_ARTWORKSOUNDS_DEST"], sound.path_artwork()), mime=True
+        )
+        obj_artwork = {"type": "Image", "url": url_artwork, "mediaType": mime_artwork}
     else:
-        url_artwork = None
+        obj_artwork = None
 
     # Fetch object and update fields
-    object = sound.activity.payload["object"]
+    obj = sound.activity.payload["object"]
     object["name"] = sound.title
     object["content"] = sound.description
     # custom things that can change
@@ -711,11 +725,11 @@ def send_update_sound(sound: Sound) -> None:
     licence = sound.licence_info()
     licence["id"] = str(licence["id"])  # integers makes jsonld cry
     object["licence"] = licence
-    object["artwork"] = url_artwork
+    object["image"] = obj_artwork
 
     to = [follower.actor.url for follower in actor.followers]
     to.append(ap.AS_PUBLIC)
-    raw_update = dict(to=to, actor=actor.to_dict(), object=object)
+    raw_update = dict(to=to, actor=actor.to_dict(), object=obj)
     raw_update["@context"] = DEFAULT_CTX
     current_app.logger.debug(f"recipients: {raw_update['to']}")
     update = ap.Update(**raw_update)